Comment posting
Feature complete. I first tried making Post a hidden field with the value passed by URL to the form, instead of a select like Author, which doesn't make sense, but that didn't seem to just work, so this'll have to do. At least the proper Post is selected by default.