two column layout
I used the bootstrap grid system to create a two column layout for the
blog posts on the main page. On very narrow screens, they are still
displayed in one column. I changed the css so that only one line of
text is shown for each blog post, because different heights would turn
the layout into chaos.